+++ cpp/subprojects/boosting/src/boosting/binning/label_binning_equal_width.cpp
@@ -1,6 +1,8 @@
 #include "boosting/binning/label_binning_equal_width.hpp"
 #include "boosting/rule_evaluation/rule_evaluation_example_wise_complete_binned.hpp"
+#include "boosting/rule_evaluation/rule_evaluation_example_wise_partial_fixed_binned.hpp"
 #include "boosting/rule_evaluation/rule_evaluation_label_wise_complete_binned.hpp"
+#include "boosting/rule_evaluation/rule_evaluation_label_wise_partial_fixed_binned.hpp"
 #include "common/math/math.hpp"
 #include "common/util/validation.hpp"
 #include <limits>
@@ -214,6 +216,17 @@ namespace boosting {
             l1RegularizationWeight, l2RegularizationWeight, std::move(labelBinningFactoryPtr));
     }
 
+    std::unique_ptr<ILabelWiseRuleEvaluationFactory> EqualWidthLabelBinningConfig::createLabelWiseFixedPartialRuleEvaluationFactory(
+            float32 labelRatio, uint32 minLabels, uint32 maxLabels) const {
+        float64 l1RegularizationWeight = l1RegularizationConfigPtr_->getWeight();
+        float64 l2RegularizationWeight = l2RegularizationConfigPtr_->getWeight();
+        std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
+            std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
+        return std::make_unique<LabelWiseFixedPartialBinnedRuleEvaluationFactory>(
+            labelRatio, minLabels, maxLabels, l1RegularizationWeight, l2RegularizationWeight,
+            std::move(labelBinningFactoryPtr));
+    }
+
     std::unique_ptr<IExampleWiseRuleEvaluationFactory> EqualWidthLabelBinningConfig::createExampleWiseCompleteRuleEvaluationFactory(
             const Blas& blas, const Lapack& lapack) const {
         float64 l1RegularizationWeight = l1RegularizationConfigPtr_->getWeight();
@@ -222,6 +235,17 @@ namespace boosting {
             std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
         return std::make_unique<ExampleWiseCompleteBinnedRuleEvaluationFactory>(
             l1RegularizationWeight, l2RegularizationWeight, std::move(labelBinningFactoryPtr), blas, lapack);
+    }
+
+    std::unique_ptr<IExampleWiseRuleEvaluationFactory> EqualWidthLabelBinningConfig::createExampleWiseFixedPartialRuleEvaluationFactory(
+            float32 labelRatio, uint32 minLabels, uint32 maxLabels, const Blas& blas, const Lapack& lapack) const {
+        float64 l1RegularizationWeight = l1RegularizationConfigPtr_->getWeight();
+        float64 l2RegularizationWeight = l2RegularizationConfigPtr_->getWeight();
+        std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr =
+            std::make_unique<EqualWidthLabelBinningFactory>(binRatio_, minBins_, maxBins_);
+        return std::make_unique<ExampleWiseFixedPartialBinnedRuleEvaluationFactory>(
+            labelRatio, minLabels, maxLabels, l1RegularizationWeight, l2RegularizationWeight,
+            std::move(labelBinningFactoryPtr), blas, lapack);
     }
 
 }